Fix missing matches in the call hierarchy
Integrate the annotations in the hierarchy using their own set of
navigated types. The cycle-detection algorithm assumed that the mere
presence of a type in the set of visited types was enough to detect a
cycle in the hierarchy of classes/interfaces, but integrating the
annotations can navigate down the hierarchy (i.e. to subclasses) leading
to false positives in the algorithm.